ARM64/dts: qcom: Disable CoreSight DT entries for waipio/cape
Made by ARM themselves, CoreSight is a set of debug and tracing drivers that
are used on ARM SoCs. The tracing features this driver set provides are very
useful during SoC bring ups, however, come with a significantly heavy overhead.

CoreSight tracing configs are enabled by default on perf-defconfigs but it's
recommended to be disabled on production devices as per 80-PJ035-1 article due
its aforementioned heavy overhead. Furthermore, even when these debugging configs
disabled, due to these DT entries, clocks are being left on which in turn introduces
power regressions.

This commit marks the JTAG driver nodes as disabled and removes the inclusion of
the coresight devicetree which contain its properties and mappings. Finally, this
also remove other nodes that were reduntant after this change.

Signed-off-by: Daniel Jacob Chittoor <djchittoor47@gmail.com>
Signed-off-by: Sourajit Karmakar <sourajit.karmakar@gmail.com>